+++ src/autonomous_trading_platform/storage/sor/models/runtime_job_runs.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Any
 
 import sqlalchemy as sa
-from sqlalchemy import ForeignKey, Integer, String
+from sqlalchemy import BigInteger, ForeignKey, String
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -35,7 +35,7 @@ class RuntimeJobRuns(Base):
         nullable=True,
     )
 
-    duration_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    duration_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     error_message: Mapped[str | None] = mapped_column(sa.Text(), nullable=True)
     correlation_id: Mapped[str | None] = mapped_column(String(128), nullable=True)
